% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.Pseudobulk.R
\name{Data.Pseudobulk}
\alias{Data.Pseudobulk}
\title{Data Preparation for Pseudo-bulk Analysis}
\usage{
Data.Pseudobulk(
  i,
  object,
  ident.1 = "disease",
  ident.2 = "control",
  subject_ID,
  logfc.threshold = 0,
  min.pct = 0,
  min.num_subjects = 0,
  min.unique_counts = 0,
  verbose = TRUE,
  group_label,
  bulk_fun,
  fixed_effects
)
}
\arguments{
\item{i}{Indicator of cell type.}

\item{object}{A Seurat object containing the RNA data.}

\item{ident.1}{The first group identifier for comparison (default: "disease").}

\item{ident.2}{The second group identifier for comparison (default: "control").}

\item{subject_ID}{Subject ID used for identifying subjects.}

\item{logfc.threshold}{The threshold for the log fold change to filter genes.}

\item{min.pct}{The minimum detection percentage required to keep a gene.}

\item{min.num_subjects}{The minimum number of subjects required to keep a gene.}

\item{min.unique_counts}{The minimum number of unique counts required to keep a gene.}

\item{verbose}{Whether to show progress (default: TRUE).}

\item{group_label}{Label of the group used for identifying cells.}

\item{bulk_fun}{Function to use for bulk calculation ("mean" or "sum").}

\item{fixed_effects}{Fixed effects (other than "group") in the model, e.g., c("batch", "year").}
}
\value{
A list containing:
\item{cellbulk_degs}{A matrix of bulk-calculated differential expression data.}
\item{gene_stats}{A data frame containing statistics of the filtered genes.}
\item{metadata_subset_distinct}{A data frame with distinct metadata for subjects.}
}
\description{
This function processes a Seurat object to prepare data for differential expression analysis using the Pseudo-bulk method.
It filters the data and handles bulk calculation methods (mean or sum).
}
