% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeR.R
\name{DEs.edgeR}
\alias{DEs.edgeR}
\title{Differential Expression Analysis Using edgeR}
\usage{
DEs.edgeR(
  cellbulk_degs,
  gene_stats_degs,
  metadata_subset_distinct,
  subject_ID,
  bulk_fun,
  de_type = "LRT",
  normalization
)
}
\arguments{
\item{cellbulk_degs}{A matrix of bulk-calculated differential expression data.}

\item{gene_stats_degs}{A data frame containing statistics of the filtered genes.}

\item{metadata_subset_distinct}{A data frame with distinct metadata for subjects.}

\item{subject_ID}{Subject ID used for identifying subjects.}

\item{bulk_fun}{Function used for bulk calculation ("mean" or "sum").}

\item{de_type}{Type of differential expression test to use ("LRT" or "Quasi-Likelihood").}

\item{normalization}{Whether to normalize the data (TRUE or FALSE).}
}
\value{
A data frame containing the results of the differential expression analysis.
}
\description{
This function performs differential expression analysis using the edgeR package on bulk-calculated expression data.
It supports normalization and handles both LRT and Quasi-Likelihood methods.
}
