% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMM.R
\name{DEs.GLMM}
\alias{DEs.GLMM}
\title{Differential Expression Analysis using Generalized Linear Mixed Models (GLMM)}
\usage{
DEs.GLMM(
  object,
  i,
  ident.1 = "disease",
  ident.2 = "nondisease",
  logfc.threshold,
  min.pct,
  min.num_subjects,
  min.unique_counts,
  group_label,
  subject_ID,
  n_threads = cores,
  de_method = "negbinom",
  de_type = "Wald",
  fixed_effects,
  zero_intercept = FALSE,
  zero_group = FALSE,
  zero_subject = FALSE,
  zero_fixed = NULL,
  show_progress = TRUE
)
}
\arguments{
\item{object}{A Seurat object containing the RNA data to be analyzed.}

\item{i}{Indicator of the cell type to filter for.}

\item{ident.1}{The first group identifier for comparison (default is "disease").}

\item{ident.2}{The second group identifier for comparison (default is "control").}

\item{logfc.threshold}{The threshold for the log fold change to filter genes.}

\item{min.pct}{The minimum detection percentage required to keep a gene.}

\item{min.num_subjects}{The minimum number of subjects required to keep a gene.}

\item{min.unique_counts}{The minimum number of unique counts required to keep a gene.}

\item{group_label}{The label of the group (disease vs. control or treatment vs. control).}

\item{subject_ID}{The subject ID used for identifying subjects.}

\item{n_threads}{Number of threads to use for parallel processing (default is the number of cores available).}

\item{de_method}{The method for differential expression analysis ("negbinom" for negative binomial or "tweedie" for tweedie distribution, default is "negbinom").}

\item{de_type}{The type of statistical test to use ("Wald" for Wald test or "LRT" for likelihood ratio test, default is "Wald").}

\item{fixed_effects}{Fixed effects (other than "group") to include in the model, e.g., c("batch", "sex").}

\item{zero_intercept}{Logical indicating whether to include an intercept in the zero-inflation model (default is FALSE).}

\item{zero_group}{Logical indicating whether to include "group" as a fixed effect in the zero-inflation model (default is FALSE).}

\item{zero_subject}{Logical indicating whether to include "subject" as a random effect in the zero-inflation model (default is FALSE).}

\item{zero_fixed}{Fixed effects (other than "group") to include in the zero-inflation part of the model.}

\item{show_progress}{Logical indicating whether to show progress of the parallel processing (default is TRUE).}
}
\value{
A data frame containing the differential expression analysis results with columns:
\item{gene}{Gene identifier.}
\item{p_val}{P-value for the differential expression test.}
\item{test_statistic}{Test statistic for the differential expression test.}
\item{odds_ratio}{Odds ratio for the differential expression.}
\item{zi_odds_ratio}{Zero-inflation model odds ratio (if applicable).}
\item{zi_p_val}{P-value for the zero-inflation model (if applicable).}
\item{converged}{Logical indicating whether the model converged.}
\item{de_method}{Method used for differential expression analysis.}
\item{de_type}{Type of statistical test used.}
\item{p_val_adj}{Adjusted p-value using the Benjamini-Hochberg method.}
\item{logFC_sample}{Log fold change for the sample.}
\item{FC_sample}{Fold change for the sample.}
\item{detected_pct_group1}{Detection percentage in group 1.}
\item{detected_pct_group2}{Detection percentage in group 2.}
\item{unique_counts_group1}{Unique counts in group 1.}
\item{unique_counts_group2}{Unique counts in group 2.}
\item{num_subjects_group1}{Number of subjects in group 1.}
\item{num_subjects_group2}{Number of subjects in group 2.}
\item{prop_zeros_group1}{Proportion of zeros in group 1.}
\item{prop_zeros_group2}{Proportion of zeros in group 2.}
}
\description{
This function performs differential expression analysis on cell-level data using generalized linear mixed models (GLMM).
It supports both negative binomial and tweedie distributions, and offers Wald test or likelihood ratio test (LRT) options
for statistical testing. The function can also include zero-inflation models with various configurations.
}
