% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilcoxon_cluster.R
\name{DEs.wilcoxon.clustered}
\alias{DEs.wilcoxon.clustered}
\title{Differential Expression Analysis Using Wilcoxon Clustered Test}
\usage{
DEs.wilcoxon.clustered(
  i,
  object,
  ident.1 = "disease",
  ident.2 = "control",
  subject_ID,
  logfc.threshold = 0,
  min.pct = 0,
  min.num_subjects = 0,
  min.unique_counts = 0,
  test.use = "clusWilcox",
  verbose = TRUE,
  parallel = FALSE,
  group_label
)
}
\arguments{
\item{i}{Indicator of cell type.}

\item{object}{A Seurat object containing the RNA data.}

\item{ident.1}{The first group identifier for comparison (default: "disease").}

\item{ident.2}{The second group identifier for comparison (default: "control").}

\item{subject_ID}{Subject ID used for identifying subjects.}

\item{logfc.threshold}{The threshold for the log fold change to filter genes.}

\item{min.pct}{The minimum detection percentage required to keep a gene.}

\item{min.num_subjects}{The minimum number of subjects required to keep a gene.}

\item{min.unique_counts}{The minimum number of unique counts required to keep a gene.}

\item{test.use}{Statistical test to use (default: "clusWilcox").}

\item{verbose}{Whether to show progress (default: TRUE).}

\item{parallel}{Whether to run the function in parallel (default: FALSE).}

\item{group_label}{Label of the group used for identifying cells.}
}
\value{
A data frame containing the results of the differential expression analysis.
}
\description{
This function performs differential expression analysis using the Wilcoxon clustered test on cell-level data.
It is designed to compare two groups within a given cell type, accounting for clustering by subject.
}
