% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wilcoxon.R
\name{DEs.wilcoxon}
\alias{DEs.wilcoxon}
\title{Differential Expression Analysis Using Wilcoxon rank-sum Test}
\usage{
DEs.wilcoxon(
  i,
  object,
  ident.1 = "disease",
  ident.2 = "control",
  min.cells.feature = 0,
  min.cells.group = 0,
  logfc.threshold = 0,
  min.pct = 0,
  test.use = "wilcox",
  verbose = TRUE,
  group_label
)
}
\arguments{
\item{i}{Indicator of cell type.}

\item{object}{A Seurat object containing the RNA data.}

\item{ident.1}{The first group identifier for comparison (default: "disease").}

\item{ident.2}{The second group identifier for comparison (default: "control").}

\item{min.cells.feature}{Minimum number of cells expressing a feature to include (default: 0).}

\item{min.cells.group}{Minimum number of cells in each group to include a feature (default: 0).}

\item{logfc.threshold}{The threshold for the log fold change to filter genes (default: 0).}

\item{min.pct}{Minimum detection percentage required to keep a gene (default: 0).}

\item{test.use}{Statistical method to use (default: "wilcox").}

\item{verbose}{Whether to show progress (default: TRUE).}

\item{group_label}{The label of the group (disease vs. control or treatment vs. control).}
}
\value{
A data frame containing the results of the differential expression analysis.
}
\description{
This function performs differential expression analysis using the Wilcoxon rank-sum test on cell-level data.
It is designed to compare two groups within a given cell type.
}
