% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene_filter.R
\name{Gene_filter}
\alias{Gene_filter}
\title{Gene Filter Function}
\usage{
Gene_filter(
  object,
  i,
  ident.1,
  ident.2,
  logfc.threshold,
  min.pct,
  min.num_subjects,
  min.unique_counts,
  group_label,
  subject_ID
)
}
\arguments{
\item{object}{A Seurat object containing the RNA data to be filtered.}

\item{i}{The cell type to filter for.}

\item{ident.1}{The first group identifier for comparison (disease).}

\item{ident.2}{The second group identifier for comparison (control).}

\item{logfc.threshold}{The threshold for the log fold change to filter genes.}

\item{min.pct}{The minimum detection percentage required to keep a gene.}

\item{min.num_subjects}{The minimum number of subjects required to keep a gene.}

\item{min.unique_counts}{The minimum number of unique counts required to keep a gene.}

\item{group_label}{The label of the group (disease vs. control or treatment vs. control).}

\item{subject_ID}{The subject ID used for identifying subjects.}
}
\value{
A list containing:
\item{object_filtered}{A Seurat object with the filtered genes.}
\item{gene_stats}{A data frame containing statistics of the filtered genes.}
}
\description{
This function filters genes in a Seurat object based on various criteria such as log fold change, detection percentage, minimum number of subjects
and unique counts per subject. It returns a filtered Seurat object and a data frame with gene statistics.
}
